refactor: centralize pitch spelling
