+++ src/lib/musicxml-export.test.ts
@@ -204,35 +204,6 @@ describe("MusicXML export", () => {
     });
   });
 
-  it("derives flat note spelling in the export model", () => {
-    const model = buildModel([makeNote({ pitch: 34 })], {
-      keySignature: { fifths: -2, mode: "minor" },
-    });
-
-    expect(model.measures[0][0]).toMatchObject({
-      type: "note",
-      pitch: { step: "B", alter: -1, octave: 2 },
-    });
-  });
-
-  it("derives altered natural spelling from the exact key signature", () => {
-    const flatModel = buildModel([makeNote({ pitch: 47 })], {
-      keySignature: { fifths: -6, mode: "minor" }, // Eb minor
-    });
-    const sharpModel = buildModel([makeNote({ pitch: 41 })], {
-      keySignature: { fifths: 6, mode: "minor" }, // D# minor
-    });
-
-    expect(flatModel.measures[0][0]).toMatchObject({
-      type: "note",
-      pitch: { step: "C", alter: -1, octave: 4 },
-    });
-    expect(sharpModel.measures[0][0]).toMatchObject({
-      type: "note",
-      pitch: { step: "E", alter: 1, octave: 3 },
-    });
-  });
-
   it("rejects a polyphonic chord", () => {
     expect(() =>
       buildModel([
