+++ src/lib/musicxml-export.ts
@@ -1,4 +1,9 @@
-import type { KeySignature, Note, TimeSignature } from "../types";
+import type { Note, TimeSignature } from "../types";
+import {
+  type KeySignature,
+  type SpelledPitch,
+  spellMidiPitch,
+} from "./pitch-spelling";
 import { resolveTabPosition, type TabPosition } from "./tab-annotation";
 
 // This format was manually reduced from a MuseScore MusicXML export and
@@ -35,19 +40,13 @@ export type MusicXmlMeasureEvent =
   | { type: "rest"; duration: number; notation: DurationNotation }
   | {
       type: "note";
-      pitch: MusicXmlPitch;
+      pitch: SpelledPitch;
       duration: number;
       notation: DurationNotation;
       tabPosition: TabPosition;
       tieStart: boolean;
       tieStop: boolean;
     };
-
-type MusicXmlPitch = {
-  step: string;
-  alter: number;
-  octave: number;
-};
 
 type DurationNotation = {
   type: string;
@@ -221,7 +220,7 @@ function buildMeasureEvents({
       const pieceEnd = pieceStart + piece.duration;
       events.push({
         type: "note",
-        pitch: midiPitchToMusicXml(note.note.pitch, keySignature),
+        pitch: spellMidiPitch({ pitch: note.note.pitch, keySignature }),
         duration: piece.duration,
         notation: piece.notation,
         tabPosition: note.tabPosition,
@@ -444,7 +443,7 @@ function renderStaffDetails(
     { number: 2 },
     hx("staff-lines", openStringPitches.length),
     ...tuning.map((midi, index) => {
-      const pitch = midiPitchToMusicXml(midi, keySignature);
+      const pitch = spellMidiPitch({ pitch: midi, keySignature });
       return h(
         "staff-tuning",
         { line: index + 1 },
@@ -511,93 +510,6 @@ function renderDurationNotation(notation: DurationNotation): XmlNode[] {
     notation.triplet &&
       hx("time-modification", hx("actual-notes", 3), hx("normal-notes", 2)),
   ];
-}
-
-function midiPitchToMusicXml(
-  pitch: number,
-  keySignature: KeySignature,
-): MusicXmlPitch {
-  const naturalPitchClasses = {
-    C: 0,
-    D: 2,
-    E: 4,
-    F: 5,
-    G: 7,
-    A: 9,
-    B: 11,
-  } as const;
-  const accidentalOrder =
-    keySignature.fifths < 0
-      ? (["B", "E", "A", "D", "G", "C", "F"] as const)
-      : (["F", "C", "G", "D", "A", "E", "B"] as const);
-  const keyAlter = Math.sign(keySignature.fifths);
-  const alteredSteps = new Set(
-    accidentalOrder.slice(0, Math.abs(keySignature.fifths)),
-  );
-  const pitchClass = pitch % 12;
-  for (const [step, naturalPitchClass] of Object.entries(naturalPitchClasses)) {
-    const alter = alteredSteps.has(step as keyof typeof naturalPitchClasses)
-      ? keyAlter
-      : 0;
-    if ((naturalPitchClass + alter + 12) % 12 === pitchClass) {
-      return {
-        step,
-        alter,
-        octave: Math.floor((pitch - naturalPitchClass - alter) / 12),
-      };
-    }
-  }
-
-  // Chromatic fallback:
-  //   key signature has flats    => prefer flats
-  //   key signature has no flats => prefer sharps
-  // `fifths` is MusicXML's signed count: negative for flats, positive for
-  // sharps, and zero for C major or A minor.
-  // Examples:
-  //   F major,  pitch class 10 -> Bb (correct: Bb chord)
-  //   A minor,  pitch class 8  -> G# (correct: E7)
-  //   C major,  pitch class 10 -> A# (wrong: Bb in Gm7 -> C7 -> F)
-  //   C major,  pitch class 1  -> C# (correct: A7; wrong: Db in Db7)
-  //   A minor,  pitch class 8  -> G# (correct: E7; wrong: Ab in Ab7)
-  // TODO: Prefer common chromatic spellings. Contextual cases require harmonic
-  // analysis or an explicit per-note spelling choice.
-  const sharpPitchClasses = [
-    ["C", 0],
-    ["C", 1],
-    ["D", 0],
-    ["D", 1],
-    ["E", 0],
-    ["F", 0],
-    ["F", 1],
-    ["G", 0],
-    ["G", 1],
-    ["A", 0],
-    ["A", 1],
-    ["B", 0],
-  ] as const;
-  const flatPitchClasses = [
-    ["C", 0],
-    ["D", -1],
-    ["D", 0],
-    ["E", -1],
-    ["E", 0],
-    ["F", 0],
-    ["G", -1],
-    ["G", 0],
-    ["A", -1],
-    ["A", 0],
-    ["B", -1],
-    ["B", 0],
-  ] as const;
-  const pitchClasses =
-    keySignature.fifths < 0 ? flatPitchClasses : sharpPitchClasses;
-  const [step, alter] = pitchClasses[pitch % 12];
-  const naturalPitchClass = naturalPitchClasses[step];
-  return {
-    step,
-    alter,
-    octave: Math.floor((pitch - naturalPitchClass - alter) / 12),
-  };
 }
 
 // xml hyperscript helpers
